Issue #219: Move fullscreenBackground code to BrowserViewModel verbatim.
This is roughly a verbatim move: the following commit will be to rename
and fix up inconsistencies.

I originally thought this bug could only be fixed from windowBackground.
However, in researching another issue, I discovered this was not true.
This is a better solution because it encapsulates this functionality
to the BrowserFragment and does not couple the windowBackground, which
may need to be used for other things, to this view.

Since the prior patches have already been reviewed, I'm putting this
change on top.
